requirements.txt
APScheduler==3.6.3
astroid==2.9.3
autopep8==1.6.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
DAWG-Python==0.7.2
docopt==0.6.2
et-xmlfile==1.1.0
flake8==4.0.1
google-api-core==2.4.0
google-api-python-client==2.29.0
google-auth==2.5.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
openpyxl==3.0.9
platformdirs==2.4.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.12.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.7
pyTelegramBotAPI==4.3.1
python-dotenv==0.19.2
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rsa==4.8
six==1.16.0
telebot==0.0.4
toml==0.10.2
tornado==6.1
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
